fix lỗi thiếu trường dữ liệu trong Models/VatTu.cs

+++ QuanLyVatTu_ASP/Controllers/GioHangController.cs
@@ -45,7 +45,7 @@ namespace QuanLyVatTu_ASP.Controllers
                 {
                     VatTuId = vatTu.ID,
                     TenVatTu = vatTu.TenVatTu,
-                    DonGia = (double)vatTu.GiaBan,
+                    DonGia = (decimal)vatTu.GiaBan,
                     SoLuong = quantity,
                 });
             }
